+++ docs/HISTORY.txt
@@ -4,8 +4,9 @@ Changelog
 2.5-dev0 - (unreleased)
 ---------------------------
 * remove edit user permission for nfps [dumitval]
-* move _is_authenticated, _get_ldap_agent, logged_in_user, nfp_for_country
-  orgs_in_country to logic_common [dumitval]
+* move _is_authenticated, logged_in_user  to logic_common [dumitval]
+* move _get_ldap_agent to ldap_config [dumitval]
+* move nfp_for_country, orgs_in_country to ui_common [dumitval]
 * _get_ldap_agent default bind=True [dumitval]
 * removed unused code for auto-disabling users [dumitval]
