+++ eea/ldapadmin/ui_common.py
@@ -1,13 +1,15 @@
 ''' common ui methods '''
+import six
 from zope.component import getMultiAdapter
 from Acquisition import Implicit
 from eea.ldapadmin.constants import NETWORK_NAME
 from eea.ldapadmin import roles_leaders
-from eea.ldapadmin.countries import get_country
+from eea.ldapadmin.countries import get_country, get_country_options
 from eea.ldapadmin.logic_common import logged_in_user, _is_authenticated
+from eea.ldapadmin.logic_common import load_template
+from eea.ldapadmin.ldap_config import _get_ldap_agent
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from Products.PageTemplates.PageTemplateFile import PageTemplateFile
-from z3c.pt.pagetemplate import PageTemplateFile as ChameleonTemplate
 
 
 def get_role_name(agent, role_id):
@@ -52,21 +54,6 @@ def extend_crumbs(crumbs_html, editor_url, extra_crumbs):
     last_crumb.text = last_crumb_text
 
     return tostring(crumbs)
-
-
-# pylint: disable=dangerous-default-value
-def load_template(name, context=None, _memo={}):
-    ''' load the main template '''
-    if name not in _memo:
-        tpl = ChameleonTemplate(name)
-
-        if context is not None:
-            bound = tpl.bind(context)
-            _memo[name] = bound
-        else:
-            _memo[name] = tpl
-
-    return _memo[name]
 
 
 plone5_wrapper = PageTemplateFile('zpt/plone5_wrapper.zpt', globals())
@@ -262,3 +249,46 @@ class NaayaViewPageTemplateFile(ViewPageTemplateFile):
         result = renderer.wrap(s)
 
         return result
+
+
+def orgs_in_country(context, country):
+    """ return a dict of organisations in countrys """
+    agent = _get_ldap_agent(context, secondary=True)
+    orgs_by_id = agent.all_organisations()
+    countries = dict(get_country_options(country=country))
+    orgs = {}
+
+    for org_id, info in six.iteritems(orgs_by_id):
+        country_info = countries.get(info['country'])
+
+        if country_info:
+            orgs[org_id] = info
+
+    return orgs
+
+
+def nfp_for_country(context):
+    """ Return country code for which the current user has NFP role
+        or None otherwise"""
+    user_id = context.REQUEST.AUTHENTICATED_USER.getId()
+
+    if user_id:
+        ldap_groups = get_ldap_user_groups(context, user_id)
+
+        for group in ldap_groups:
+            if ('eionet-nfp-mc-' in group[0] or
+                'eionet-nfp-cc-' in group[0] or
+                    'eionet-nfp-oc-' in group[0]):
+
+                return group[0].rsplit('-', 1)[-1]
+    return None
+
+
+def get_ldap_user_groups(context, user_id):
+    """ return the ldap roles the user is member of """
+    agent = _get_ldap_agent(context, secondary=True)
+    ldap_roles = sorted(agent.member_roles_info('user',
+                                                user_id,
+                                                ('description',)))
+
+    return ldap_roles
